refactor(integration): decompose meta_analysis/visualizer.py into mixin siblings
MAJ-04 module 5/6: integration/meta_analysis/visualizer.py 2871 -> 283
lines (facade).

Class-method bodies must move verbatim to stay byte-identical, so the
plot methods move into Sweep*Mixin siblings (the execute/processor.py
pattern adapted for a single-class module):

- visualizer_style.py: matplotlib bootstrap (single owner of the
  matplotlib.use('Agg') side effect and _MPL_AVAILABLE), _STYLE,
  _FRAMEWORK_COLORS, _get_color, _fmt_time, _add_watermark
- visualizer_export.py: SweepExportMixin (_export_csv)
- visualizer_runtime_plots.py: SweepRuntimePlotMixin (9 plot methods)
- visualizer_metric_plots.py: SweepMetricPlotMixin (6 plot methods)
- visualizer_summary_plots.py: SweepSummaryPlotMixin (5 plot methods)
- visualizer.py: facade - SweepVisualizer(mixins) keeps __init__,
  generate_all, and _safe_log_scale verbatim; re-exports every
  module-level name incl. plt/np/mcolors/mticker/Axes3D

The test's monkeypatch target (visualizer._MPL_AVAILABLE,
tests/integration/test_integration_meta_analysis_validation.py:287-289)
stays on the facade and generate_all still reads the facade binding, so
no test edits. Mixins carry TYPE_CHECKING attribute stubs (records/
output_dir/logger/_gnn_format_statistics/_safe_log_scale) preserving the
original typing context for mypy. Mixin headers import matplotlib names
from visualizer_style; matplotlib is a core dependency, and the
defensive try/except bootstrap remains the single owner in style.

2811 lines verified byte-identical (method blocks incl. section banners
+ kept facade methods + style chunks); only import headers, the class
base list, and TYPE_CHECKING stubs differ.

Verified: ruff 0 (repo-wide); mypy 0; facade contract 13 module names +
SweepVisualizer surface; targeted pytest
test_integration_meta_analysis_validation.py: 14 passed (incl. the
_MPL_AVAILABLE monkeypatch test).
